refactor(codegen): extract std::io::tcp::* into stdlib/io_tcp.rs (Round 1.13)
13 fns lifted into IoTcpStdlib<'ctx>: listen_socket{,_fallible},
connect{,_fallible}, accept_one{,_fallible}, send, send_bytes, recv,
recv_bytes, recv_into, close_fd, shutdown_listen_socket. Covers both
the older non-fallible variants and the IoError-fallible versions
that ship in v1.x.

Surface bumps to pub(crate):
- Cx method: global_string (used to emit string-pointer literals)

codegen.rs: 41625 -> 40745 (-880 lines).
